\name{validateView}
\alias{validateView}
\title{Checks views defined in \code{viewSpec}}
\usage{
validateView(vs)
}
\arguments{
  \item{vs}{object returned by viewSpec.}
}
\value{
NULL (Throws an error or prints reassurance depending on
correctness).
}
\description{
\code{validateView} checks views against table fields and
upstream views for
}

